stance detection datasets added
added 3 new stance detection datasets:
Perspectrum - contains a claim and perspective. There are various perspectives to the same claim, and they can oppose or support it.
SemEval2016 - contains tweets, topics and the stance of the tweet on the topic (oppose or support)
Xstance - contains political questions and comments with the label is the stance of the comment towards the question. contains questions in French, Italian and German.